requirements.txt
aiohttp
websockets
sqlalchemy
scipy
numpy
pandas
scikit-learn
xgboost
pyyaml
matplotlib
